requirements.txt
streamlit
pandas
psycopg2-binary
bcrypt
streamlit-drawable-canvas
streamlit-folium
